% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmAlignment.R
\name{alignPwmSets}
\alias{alignPwmSets}
\title{Multiple PWMs alignment}
\usage{
alignPwmSets(left_pwms_set, left_alignment, right_pwms_set,
  right_alignment, try_reverse_complement)
}
\arguments{
\item{left_pwms_set}{list of pwms(matrixes)}

\item{left_alignment}{alignment of left_pwms_set.}

\item{right_pwms_set}{list of pwms;}

\item{right_alignment}{alignment of right_pwms_set.}

\item{try_reverse_complement}{if true(default), also try reverse complement.}
}
\value{
list - alignment of concatination of left_pwms_set and right_pwms_set
}
\description{
Align two sets of pwms
}
\author{
Lando Andrey
}
