\name{Alphabet}
\alias{Alphabet}
\title{built alphabet}
\usage{
  Alphabet(chars, cols)
}
\arguments{
  \item{chars}{set of symbols}

  \item{cols}{set of colors; one for each symbol}
}
\value{
  the Alphabet object
}
\description{
  builts an object of class Alphabet from the given set of
  symbols and colors
}
\author{
  Martin Nettling
}

