% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSeqLogoSupport.R
\name{baseDistributionPwm}
\alias{baseDistributionPwm}
\title{Generates a PWM consisting of only the uniform distribtuion or the given base_distribution (if defined).}
\usage{
baseDistributionPwm(pwm_length, alphabet_length, base_distribution = NULL)
}
\description{
Generates a PWM consisting of only the uniform distribtuion or the given base_distribution (if defined).
}
