\name{diffLogo}
\alias{diffLogo}
\title{Draw DiffLogo}
\usage{
  diffLogo(diffLogoObj, ymin = 0, ymax = 0, sparse = FALSE)
}
\arguments{
  \item{diffLogoObj}{a DiffLogoObject created by the
  function createDiffLogoObject}

  \item{ymin}{minimum value on the y-axis}

  \item{ymax}{maximum value on the y-axis}

  \item{sparse}{if TRUE margins are reduced and tickmarks
  are removed from the logo}
}
\description{
  Draws the difference of two sequence logos.
}
\author{
  Martin Nettling
}

