% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{FULL_ALPHABET}
\alias{FULL_ALPHABET}
\title{Complete character alphabet}
\format{An object of class \code{Alphabet} of length 4.}
\usage{
FULL_ALPHABET
}
\description{
the alphabet of all 26 characters
}
\examples{
FULL_ALPHABET
}
\author{
Hendrik Treutler
}
\keyword{datasets}
