% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getAlphabetFromCharacters}
\alias{getAlphabetFromCharacters}
\title{returns the alphabet which fits to the given characters}
\usage{
getAlphabetFromCharacters(characters)
}
\arguments{
\item{characters}{a character vector of characters}
}
\value{
an alphabet of type Alphabet
}
\description{
returns the alphabet which fits to the given characters
}
\examples{
alphabet = getAlphabetFromSequences(c("A","A","C","C","G","G","T","T"))
}
