% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getAlphabetFromSequences}
\alias{getAlphabetFromSequences}
\title{returns the alphabet which fits to the given sequences}
\usage{
getAlphabetFromSequences(sequences)
}
\arguments{
\item{sequences}{a character vector of sequences}
}
\value{
an alphabet of type Alphabet
}
\description{
returns the alphabet which fits to the given sequences
}
\examples{
alphabet = getAlphabetFromSequences("AACCGGTT")
}
