% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromAlignmentFile}
\alias{getPwmFromAlignmentFile}
\title{generates a pwm from an alignment file}
\usage{
getPwmFromAlignmentFile(filename, alphabet = NULL)
}
\arguments{
\item{filename}{the alignment file}

\item{alphabet}{the desired alphabet of type Alphabet}
}
\value{
a pwm
}
\description{
generates a pwm from an alignment file
}
\examples{
fileName = "extdata/alignments/calamodulin_1.txt"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromAlignmentFile(file)
}
