% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromFastaFile}
\alias{getPwmFromFastaFile}
\title{generates a pwm from a FASTA file}
\usage{
getPwmFromFastaFile(filename, alphabet = NULL)
}
\arguments{
\item{filename}{the FASTA file}

\item{alphabet}{the desired alphabet of type Alphabet}
}
\value{
a pwm
}
\description{
generates a pwm from a FASTA file
}
\examples{
fileName = "extdata/alignments/F-box_bacteria.seq.fa"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromFastaFile(file)
}
