% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromFile}
\alias{getPwmFromFile}
\title{generates a pwm from a file of different formats}
\usage{
getPwmFromFile(filename)
}
\arguments{
\item{filename}{the file}
}
\value{
a pwm
}
\description{
Generates a pwm from a file of different formats. 
Supported formats are FASTA files (.fa, .fasta), alignment files (.txt, .text, .al, .alignment), PWM files (.pwm), JASPAR / Position Frequency Matrix files (.pfm), and homer files (.motif).
}
\examples{
fileName = "extdata/pwm/H1-hESC.pwm"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromFile(file)
}
