% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromHomerFile}
\alias{getPwmFromHomerFile}
\title{generates a pwm from a homer file}
\usage{
getPwmFromHomerFile(filename)
}
\arguments{
\item{filename}{the homer file}
}
\value{
a pwm
}
\description{
generates a pwm from a homer file
}
\examples{
fileName = "extdata/homer/CTCF_Zf_CD4.motif"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromHomerFile(file)
}
