% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromPfmOrJasparFile}
\alias{getPwmFromPfmOrJasparFile}
\title{generates a pwm from a jaspar file}
\usage{
getPwmFromPfmOrJasparFile(filename)
}
\arguments{
\item{filename}{the jaspar file}
}
\value{
a pwm
}
\description{
generates a pwm from a jaspar file
}
\examples{
fileName = "extdata/pfm/ctcf_jaspar.pfm"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromPfmOrJasparFile(file)
}
