% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPwmFromPwmFile}
\alias{getPwmFromPwmFile}
\title{generates a pwm from a pwm file}
\usage{
getPwmFromPwmFile(filename)
}
\arguments{
\item{filename}{the pwm file}
}
\value{
a pwm
}
\description{
generates a pwm from a pwm file
}
\examples{
fileName = "extdata/pwm/H1-hESC.pwm"
file = system.file(fileName, package = "DiffLogo")
pwm = getPwmFromPwmFile(file)
}
