% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSequencesFromAlignmentFile}
\alias{getSequencesFromAlignmentFile}
\title{extracts the sequences from an alignment file}
\usage{
getSequencesFromAlignmentFile(filename)
}
\arguments{
\item{filename}{the alignment file}
}
\value{
a vector of sequences
}
\description{
extracts the sequences from an alignment file
}
\examples{
fileName = "extdata/alignments/calamodulin_1.txt"
file = system.file(fileName, package = "DiffLogo")
sequences = getSequencesFromAlignmentFile(file)
}
