% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSequencesFromFastaFile}
\alias{getSequencesFromFastaFile}
\title{extracts the sequences from a FASTA file}
\usage{
getSequencesFromFastaFile(filename)
}
\arguments{
\item{filename}{the FASTA file}
}
\value{
a vector of sequences
}
\description{
extracts the sequences from a FASTA file
}
\examples{
fileName = "extdata/alignments/F-box_bacteria.seq.fa"
file = system.file(fileName, package = "DiffLogo")
sequences = getSequencesFromFastaFile(file)
}
