\name{informationContent}
\alias{informationContent}
\title{information content}
\usage{
  informationContent(p)
}
\arguments{
  \item{p}{probability vector representing the symbol
  distribution}
}
\value{
  an object consisting of height a ylab
}
\description{
  the information content for the given probability vector
}
\author{
  Martin Nettling
}

