\name{normalizedDifferenceOfProbabilities}
\alias{normalizedDifferenceOfProbabilities}
\title{normalized probability differences}
\usage{
  normalizedDifferenceOfProbabilities(p1, p2)
}
\arguments{
  \item{p1}{probability vector representing the first
  symbol distribution}

  \item{p2}{probability vector representing the second
  symbol distribution}
}
\value{
  a vector with one result for each symbol
}
\description{
  probability differences normalized by the sum of absolute
  probability differences for the given pair of probability
  vectors
}
\author{
  Martin Nettling
}

