% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{normalizePWM}
\alias{normalizePWM}
\title{normalizes the given pwm}
\usage{
normalizePWM(pwm)
}
\arguments{
\item{pwm}{a pwm}
}
\value{
a normalized pwm
}
\description{
normalizes the given pwm to column-sums of 1.0
}
\examples{
pwm = matrix(1:40, nrow = 4, dimnames = list(c("A","C","G","T"), 1:10))
pwm = normalizePWM(pwm)
}
