\name{prepareDiffLogoTable}
\alias{prepareDiffLogoTable}
\title{Prepares a DiffLogoTable and generates an object that contains the hirarchical clustering and a matrix of prepared difference logos.}
\usage{
  prepareDiffLogoTable(PWMs, alphabet = DNA,
    configuration = list(), ...)
}
\description{
  Prepares a DiffLogoTable and generates an object that
  contains the hirarchical clustering and a matrix of prepared
  difference logos.
}

