% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/baseDistrs.R
\name{probabilities}
\alias{probabilities}
\title{probabilities}
\usage{
probabilities(p)
}
\arguments{
\item{p}{probability vector representing the symbol distribution}
}
\value{
the given vector
}
\description{
the given probabilities
}
\examples{
motif_folder = "pwm"
motif_names = c("HepG2","MCF7","HUVEC","ProgFib")
motifs = list()
for (name in motif_names) {
  fileName = paste(motif_folder,"/",name,".txt",sep="")
  file = system.file(fileName, package = "DiffLogo")
  motifs[[name]] = as.matrix(read.delim(file,header=FALSE))
}

pwm1 = motifs[[motif_names[[1]]]]
pwm2 = motifs[[motif_names[[2]]]]

diffLogoFromPwm(pwm1 = pwm1, pwm2 = pwm2, baseDistribution = probabilities)
}
\author{
Martin Nettling
}

