\docType{data}
\name{RNA}
\alias{RNA}
\title{RNA alphabet}
\format{List of 4
 $ chars                   : chr [1:4] "A" "C" "G" "U"
 $ cols                    : chr [1:4] "green4" "blue" "orange" "red"
 $ size                    : int 4
 $ supportReverseComplement: logi TRUE
 - attr(*, "class")= chr "Alphabet"}
\usage{
  RNA
}
\description{
  the RNA alphabet, i.e. A, C, G, U
}
\author{
  Martin Nettling
}
\keyword{datasets}

