\name{shannonDivergence}
\alias{shannonDivergence}
\title{shannon divergence}
\usage{
  shannonDivergence(p1, p2)
}
\arguments{
  \item{p1}{probability vector representing the first
  symbol distribution}

  \item{p2}{probability vector representing the second
  symbol distribution}
}
\value{
  an object consisting of height and ylab
}
\description{
  the shannon divergence for the given pair of probability
  vectors
}
\author{
  Martin Nettling
}

