\name{sumOfAbsICDifferences}
\alias{sumOfAbsICDifferences}
\title{sum of absolute information content differences}
\usage{
  sumOfAbsICDifferences(p1, p2)
}
\arguments{
  \item{p1}{probability vector representing the first
  symbol distribution}

  \item{p2}{probability vector representing the second
  symbol distribution}
}
\value{
  an object consisting of height and ylab
}
\description{
  the sum of absolute information content differences for
  the given pair of probability vectors
}
\author{
  Martin Nettling
}

