% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stackHeights.R
\name{sumProbabilities}
\alias{sumProbabilities}
\title{sum of probabilities, i.e. 1.0}
\usage{
sumProbabilities(p)
}
\arguments{
\item{p}{probability vector representing the symbol distribution}
}
\value{
an object consisting of height and ylab
}
\description{
the sum of probabilities for the given probability vector, i.e. 1.0
}
\examples{
motif_folder = "pwm"
motif_names = c("HepG2","MCF7","HUVEC","ProgFib")
motifs = list()
for (name in motif_names) {
  fileName = paste(motif_folder,"/",name,".txt",sep="")
  file = system.file(fileName, package = "DiffLogo")
  motifs[[name]] = as.matrix(read.delim(file,header=FALSE))
}

pwm1 = motifs[[motif_names[[1]]]]
pwm2 = motifs[[motif_names[[2]]]]

diffLogoFromPwm(pwm1 = pwm1, pwm2 = pwm2, stackHeight = sumProbabilities)
}
\author{
Martin Nettling
}

