% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmAlignment.R
\name{switchDirection}
\alias{switchDirection}
\title{If 'forward' return 'reverse'
If 'reverse' return 'forward'}
\usage{
switchDirection(direction)
}
\description{
If 'forward' return 'reverse'
If 'reverse' return 'forward'
}
