\docType{data}
\name{ASN}
\alias{ASN}
\title{TODO}
\format{List of 3
 $ chars: chr [1:20] "A" "C" "D" "E" ...
 $ cols : chr [1:20] "Light green" "Green" "Dark green" "Dark green" ...
 $ size : int 20
 - attr(*, "class")= chr "Alphabet"}
\usage{
  ASN
}
\description{
  TODO
}
\author{
  Martin Nettling
}
\keyword{datasets}

