\docType{data}
\name{ASN}
\alias{ASN}
\title{ASN alphabet}
\format{List of 4
 $ chars                   : chr [1:20] "A" "C" "D" "E" ...
 $ cols                    : chr [1:20] "Light green" "Green" "Dark green" "Dark green" ...
 $ size                    : int 20
 $ supportReverseComplement: logi FALSE
 - attr(*, "class")= chr "Alphabet"}
\usage{
  ASN
}
\description{
  the amino acid alphabet (20 symbols), i.e. A, C, D, E, F, G,
  H, I, K, L, M, N, P, Q, R, S, T, V, W, Y
}
\author{
  Martin Nettling
}
\keyword{datasets}

