% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{ASN}
\alias{ASN}
\title{ASN alphabet}
\format{An object of class \code{Alphabet} of length 4.}
\usage{
ASN
}
\description{
the amino acid alphabet (20 symbols), i.e. A, C, D, E, F, G, H, I, K, L, M, N, P, Q, R, S, T, V, W, Y
}
\examples{
ASN
}
\author{
Martin Nettling
}
\keyword{datasets}
