% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffSeqLogoSupport.R
\name{baseDistributionPwm}
\alias{baseDistributionPwm}
\title{Generates a PWM}
\usage{
baseDistributionPwm(pwm_length, alphabet_length,
  base_distribution = NULL)
}
\arguments{
\item{pwm_length}{the number of positions}

\item{alphabet_length}{the alphabet size}

\item{base_distribution}{optional base distribution for each PWM position}
}
\value{
a PWM
}
\description{
Generates a PWM consisting of only the uniform distribution or the given base_distribution (if defined).
}
