\docType{data}
\name{DNA}
\alias{DNA}
\title{DNA alphabet}
\format{List of 4
 $ chars                   : chr [1:4] "A" "C" "G" "T"
 $ cols                    : chr [1:4] "green4" "blue" "orange" "red"
 $ size                    : int 4
 $ supportReverseComplement: logi TRUE
 - attr(*, "class")= chr "Alphabet"}
\usage{
  DNA
}
\description{
  the DNA alphabet, i.e. A, C, G, T
}
\author{
  Martin Nettling
}
\keyword{datasets}

