% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{DNA}
\alias{DNA}
\title{DNA alphabet}
\format{An object of class \code{Alphabet} of length 4.}
\usage{
DNA
}
\description{
the DNA alphabet, i.e. A, C, G, T
}
\examples{
DNA
}
\author{
Martin Nettling
}
\keyword{datasets}
