% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{RNA}
\alias{RNA}
\title{RNA alphabet}
\format{\preformatted{List of 3
 $ chars: chr [1:4] "A" "C" "G" "U"
 $ cols : chr [1:4] "green4" "blue" "orange" "red"
 $ size : int 4
 - attr(*, "class")= chr "Alphabet"
}}
\usage{
RNA
}
\description{
the RNA alphabet, i.e. A, C, G, U
}
\author{
Martin Nettling
}
\keyword{datasets}

