% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabet.R
\docType{data}
\name{RNA}
\alias{RNA}
\title{RNA alphabet}
\format{An object of class \code{Alphabet} of length 4.}
\usage{
RNA
}
\description{
the RNA alphabet, i.e. A, C, G, U
}
\examples{
RNA
}
\author{
Martin Nettling
}
\keyword{datasets}
