\name{sumProbabilities}
\alias{sumProbabilities}
\title{sum of probabilities, i.e. 1.0}
\usage{
  sumProbabilities(p)
}
\arguments{
  \item{p}{probability vector representing the symbol
  distribution}
}
\value{
  an object consisting of height and ylab
}
\description{
  the sum of probabilities for the given probability
  vector, i.e. 1.0
}
\author{
  Martin Nettling
}

