% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwmAlignment.R
\name{switchDirection}
\alias{switchDirection}
\title{Switches between 'forward' and 'reverse'}
\usage{
switchDirection(direction)
}
\arguments{
\item{direction}{either 'forward' or 'reverse'}
}
\value{
either 'reverse' or 'forward'
}
\description{
Switches between 'forward' and 'reverse'
}
